+++ backend/src/reservations.py
@@ -1,6 +1,6 @@
 import uuid
 from fastapi import HTTPException
-import firebase_config as firebase_config
+import db.firebase_config as firebase_config
 import accommodations as accommodations
 from datetime import timedelta, datetime
