+++ backend/tests/step_definitions/test_evaluate.py
@@ -1,5 +1,5 @@
 from pytest_bdd import parsers, given, when, then, scenario
-import main as main
+import main
 
 @scenario(scenario_name="enviar avaliação para reserva", feature_name="../feature/evaluate.feature")
 def test_rating_add():
