+++ backend/tests/step_definitions/test_fail_evaluate.py
@@ -1,7 +1,7 @@
 from pytest_bdd import parsers, given, when, then, scenario
 import pytest
 from fastapi import HTTPException
-import main as main
+import main
 
 @scenario(scenario_name="Nenhuma estrela foi adicionada", feature_name="../feature/evaluate.feature")
 def test_rating_0_stars():
